+++ sentry-android-core/src/main/java/io/sentry/android/core/ManifestMetadataReader.java
@@ -67,6 +67,7 @@ final class ManifestMetadataReader {
   static final String IDLE_TIMEOUT = "io.sentry.traces.idle-timeout";
 
   static final String ATTACH_SCREENSHOT = "io.sentry.attach-screenshot";
+  static final String CLIENT_REPORTS_ENABLE = "io.sentry.send-client-reports";
 
   /** ManifestMetadataReader ctor */
   private ManifestMetadataReader() {}
@@ -202,6 +203,9 @@ final class ManifestMetadataReader {
 
         options.setAttachScreenshot(
             readBool(metadata, logger, ATTACH_SCREENSHOT, options.isAttachScreenshot()));
+
+        options.setSendClientReports(
+            readBool(metadata, logger, CLIENT_REPORTS_ENABLE, options.isSendClientReports()));
 
         if (options.getTracesSampleRate() == null) {
           final Double tracesSampleRate = readDouble(metadata, logger, TRACES_SAMPLE_RATE);