+++ sentry/src/main/java/io/sentry/ExternalOptions.java
@@ -38,6 +38,7 @@ public final class ExternalOptions {
   private final @NotNull Set<Class<? extends Throwable>> ignoredExceptionsForType =
       new CopyOnWriteArraySet<>();
   private @Nullable Boolean printUncaughtStackTrace;
+  private @Nullable Boolean sendClientReports;
 
   @SuppressWarnings("unchecked")
   public static @NotNull ExternalOptions from(
@@ -55,6 +56,7 @@ public final class ExternalOptions {
     options.setTracesSampleRate(propertiesProvider.getDoubleProperty("traces-sample-rate"));
     options.setDebug(propertiesProvider.getBooleanProperty("debug"));
     options.setEnableDeduplication(propertiesProvider.getBooleanProperty("enable-deduplication"));
+    options.setSendClientReports(propertiesProvider.getBooleanProperty("send-client-reports"));
     final String maxRequestBodySize = propertiesProvider.getProperty("max-request-body-size");
     if (maxRequestBodySize != null) {
       options.setMaxRequestBodySize(
@@ -272,5 +274,13 @@ public final class ExternalOptions {
 
   public void setIdleTimeout(final @Nullable Long idleTimeout) {
     this.idleTimeout = idleTimeout;
+  }
+
+  public @Nullable Boolean getSendClientReports() {
+    return sendClientReports;
+  }
+
+  public void setSendClientReports(final @Nullable Boolean sendClientReports) {
+    this.sendClientReports = sendClientReports;
   }
 }